Remove use of login button #9
