+++ src/components/FeedList.js
@@ -50,7 +50,9 @@ export const FeedList = () => {
   return (
     <Wrapper>
       {data?.feed?.length ? (
-        data.feed.map((tweet) => <Tweet key={tweet.id} tweet={tweet} offset={data?.feed?.length} />)
+        data.feed.map((tweet) => (
+          <Tweet key={tweet.id} tweet={tweet} offset={data?.feed?.length} />
+        ))
       ) : (
         <CustomResponse text="Follow some people to get some feed updates" />
       )}
